refactor: Remove mock module
